+++ Data/Vinyl.hs
@@ -1,10 +1,11 @@
 module Data.Vinyl
   ( module Data.Vinyl.Core
+  , module Data.Vinyl.ARec
   , module Data.Vinyl.Derived
   , module Data.Vinyl.Lens
   ) where
 
 import Data.Vinyl.Core
+import Data.Vinyl.ARec (ARec, toARec, fromARec)
 import Data.Vinyl.Derived
 import Data.Vinyl.Lens
-