+++ Data/Vinyl/Derived.hs
@@ -12,10 +12,12 @@
 {-# LANGUAGE UndecidableInstances #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE FlexibleContexts #-}
-
+{-# LANGUAGE TypeApplications #-}
+-- | Commonly used 'Rec' instantiations.
 module Data.Vinyl.Derived where
 
 import Data.Proxy
+import Data.Vinyl.ARec
 import Data.Vinyl.Core
 import Data.Vinyl.Functor
 import Data.Vinyl.Lens
@@ -31,8 +33,19 @@ type a ::: b = '(a, b)
 data ElField (field :: (Symbol, *)) where
   Field :: KnownSymbol s => !t -> ElField '(s,t)
 
+-- | A record of named fields.
 type FieldRec = Rec ElField
+
+-- | An 'ARec' of named fields to provide constant-time field access.
+type AFieldRec ts = ARec ElField ts
+
+
+-- | Heterogeneous list whose elements are evaluated during list
+-- construction.
 type HList = Rec Identity
+
+-- | Heterogeneous list whose elements are left as-is during list
+-- construction (cf. 'HList').
 type LazyHList = Rec Thunk
 
 deriving instance Eq t => Eq (ElField '(s,t))
@@ -45,6 +58,7 @@ instance Show t => Show (ElField '(s,t)) where
 getField :: ElField '(s,t) -> t
 getField (Field x) = x
 
+-- | Get the label name of an 'ElField'.
 getLabel :: forall s t. ElField '(s,t) -> String
 getLabel (Field _) = symbolVal (Proxy::Proxy s)
 
@@ -60,32 +74,42 @@ rfield f (Field x) = fmap Field (f x)
 {-# INLINE rfield #-}
 
 infix 3 =:
+
+-- | Operator for creating an 'ElField'. With the @-XOverloadedLabels@
+-- extension, this permits usage such as, @#foo =: 23@ to produce a
+-- value of type @ElField ("foo" ::: Int)@.
 (=:) :: KnownSymbol l => Label (l :: Symbol) -> (v :: *) -> ElField (l ::: v)
 _ =: v = Field v
 
 -- | Get a named field from a record.
 rgetf
-  :: forall l f v us. HasField l us v
-  => Label l -> Rec f us -> f (l ::: v)
+  :: forall l f v record us. HasField record l us v
+  => Label l -> record f us -> f (l ::: v)
 rgetf _ = rget (Proxy :: Proxy (l ::: v))
 
 -- | Get the value associated with a named field from a record.
 rvalf
-  :: HasField l us v => Label l -> Rec ElField us -> v
+  :: HasField record l us v => Label l -> record ElField us -> v
 rvalf x = getField . rgetf x
 
+-- | Set a named field. @rputf #foo 23@ sets the field named @#foo@ to
+-- @23@.
+rputf :: forall l v record us. (HasField record l us v, KnownSymbol l)
+      => Label l -> v -> record ElField us -> record ElField us
+rputf _ = rput . (Field :: v -> ElField '(l,v))
+
 -- | A lens into a 'Rec' identified by a 'Label'.
-rlensf' :: forall l v g f us. (Functor g, HasField l us v)
+rlensf' :: forall l v record g f us. (Functor g, HasField record l us v)
         => Label l
         -> (f (l ::: v) -> g (f (l ::: v)))
-        -> Rec f us
-        -> g (Rec f us)
+        -> record f us
+        -> g (record f us)
 rlensf' _ f = rlens (Proxy :: Proxy (l ::: v)) f
 
 -- | A lens into the payload value of a 'Rec' field identified by a
 -- 'Label'.
-rlensf :: forall l v g f us. (Functor g, HasField l us v)
-       => Label l -> (v -> g v) -> Rec ElField us -> g (Rec ElField us)
+rlensf :: forall l v record g f us. (Functor g, HasField record l us v)
+       => Label l -> (v -> g v) -> record ElField us -> g (record ElField us)
 rlensf _ f = rlens (Proxy :: Proxy (l ::: v)) (rfield f)
 
 -- | Shorthand for a 'FieldRec' with a single field.
@@ -114,8 +138,8 @@ type family FieldType l fs where
   FieldType l ((l ::: v) ': fs) = v
   FieldType l ((l' ::: v') ': fs) = FieldType l fs
 
-type HasField l fs v =
-  (RElem (l ::: v) fs (RIndex (l ::: v) fs), FieldType l fs ~ v)
+type HasField record l fs v =
+  (RecElem record (l ::: v) fs (RIndex (l ::: v) fs), FieldType l fs ~ v)
 
 -- proxy for label type
 data Label (a :: Symbol) = Label
@@ -160,3 +184,5 @@ rlabels = rpuref getLabel'
   where getLabel' :: forall l v. KnownSymbol l
                   => Const String (l ::: v)
         getLabel' = Const (symbolVal (Proxy::Proxy l))
+
+-- * Specializations for working with an 'ARec' of named fields.