+++ Data/Vinyl/Lens.hs
@@ -7,10 +7,12 @@
 {-# LANGUAGE ScopedTypeVariables   #-}
 {-# LANGUAGE TypeFamilies          #-}
 {-# LANGUAGE TypeOperators         #-}
-
+-- | Lenses into record fields.
 module Data.Vinyl.Lens
-  ( RElem(..)
-  , RSubset(..)
+  ( RecElem(..)
+  , RElem
+  , RecSubset(..)
+  , RSubset
   , REquivalent
   , type (∈)
   , type (⊆)
@@ -28,7 +30,7 @@ import Data.Typeable (Proxy(..))
 -- The third parameter to 'RElem', @i@, is there to help the constraint solver
 -- realize that this is a decidable predicate with respect to the judgemental
 -- equality in @k@.
-class i ~ RIndex r rs => RElem (r :: k) (rs :: [k]) (i :: Nat) where
+class i ~ RIndex r rs => RecElem record (r :: k) (rs :: [k]) (i :: Nat) where
 
   -- | We can get a lens for getting and setting the value of a field which is
   -- in a record. As a convenience, we take a proxy argument to fix the
@@ -40,13 +42,13 @@ class i ~ RIndex r rs => RElem (r :: k) (rs :: [k]) (i :: Nat) where
     :: Functor g
     => sing r
     -> (f r -> g (f r))
-    -> Rec f rs
-    -> g (Rec f rs)
+    -> record f rs
+    -> g (record f rs)
 
   -- | For Vinyl users who are not using the @lens@ package, we provide a getter.
   rget
     :: sing r
-    -> Rec f rs
+    -> record f rs
     -> f r
 
   -- | For Vinyl users who are not using the @lens@ package, we also provide a
@@ -54,21 +56,24 @@ class i ~ RIndex r rs => RElem (r :: k) (rs :: [k]) (i :: Nat) where
   -- and so we do not take a proxy argument here.
   rput
     :: f r
-    -> Rec f rs
-    -> Rec f rs
+    -> record f rs
+    -> record f rs
+
+-- | 'RecElem' for classic vinyl 'Rec' types.
+type RElem = RecElem Rec
 
 -- This is an internal convenience stolen from the @lens@ library.
 lens
   :: Functor f
-  => (t -> s)
-  -> (t -> a -> b)
-  -> (s -> f a)
-  -> t
-  -> f b
+  => (s -> a)
+  -> (s -> b -> t)
+  -> (a -> f b)
+  -> s
+  -> f t
 lens sa sbt afb s = fmap (sbt s) $ afb (sa s)
 {-# INLINE lens #-}
 
-instance RElem r (r ': rs) 'Z where
+instance RecElem Rec r (r ': rs) 'Z where
   rlens _ f (x :& xs) = fmap (:& xs) (f x)
   {-# INLINE rlens #-}
   rget k = getConst . rlens k Const
@@ -76,7 +81,7 @@ instance RElem r (r ': rs) 'Z where
   rput y = getIdentity . rlens Proxy (\_ -> Identity y)
   {-# INLINE rput #-}
 
-instance (RIndex r (s ': rs) ~ 'S i, RElem r rs i) => RElem r (s ': rs) ('S i) where
+instance (RIndex r (s ': rs) ~ 'S i, RElem r rs i) => RecElem Rec r (s ': rs) ('S i) where
   rlens p f (x :& xs) = fmap (x :&) (rlens p f xs)
   {-# INLINE rlens #-}
   rget k = getConst . rlens k Const
@@ -88,38 +93,40 @@ instance (RIndex r (s ': rs) ~ 'S i, RElem r rs i) => RElem r (s ': rs) ('S i) w
 -- record to the former's is evident. That is, we can either cast a larger
 -- record to a smaller one, or we may replace the values in a slice of a
 -- record.
-class is ~ RImage rs ss => RSubset (rs :: [k]) (ss :: [k]) is where
+class is ~ RImage rs ss => RecSubset record (rs :: [k]) (ss :: [k]) is where
 
   -- | This is a lens into a slice of the larger record. Morally, we have:
   --
   -- > rsubset :: Lens' (Rec f ss) (Rec f rs)
   rsubset
     :: Functor g
-    => (Rec f rs -> g (Rec f rs))
-    -> Rec f ss
-    -> g (Rec f ss)
+    => (record f rs -> g (record f rs))
+    -> record f ss
+    -> g (record f ss)
 
   -- | The getter of the 'rsubset' lens is 'rcast', which takes a larger record
   -- to a smaller one by forgetting fields.
   rcast
-    :: Rec f ss
-    -> Rec f rs
+    :: record f ss
+    -> record f rs
   rcast = getConst . rsubset Const
   {-# INLINE rcast #-}
 
   -- | The setter of the 'rsubset' lens is 'rreplace', which allows a slice of
   -- a record to be replaced with different values.
   rreplace
-    :: Rec f rs
-    -> Rec f ss
-    -> Rec f ss
+    :: record f rs
+    -> record f ss
+    -> record f ss
   rreplace rs = getIdentity . rsubset (\_ -> Identity rs)
   {-# INLINE rreplace #-}
 
-instance RSubset '[] ss '[] where
+type RSubset = RecSubset Rec
+
+instance RecSubset Rec '[] ss '[] where
   rsubset = lens (const RNil) const
 
-instance (RElem r ss i , RSubset rs ss is) => RSubset (r ': rs) ss (i ': is) where
+instance (RElem r ss i , RSubset rs ss is) => RecSubset Rec (r ': rs) ss (i ': is) where
   rsubset = lens (\ss -> rget Proxy ss :& rcast ss) set
     where
       set :: Rec f ss -> Rec f (r ': rs) -> Rec f ss